% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plotNASCByBeamAndLayer}
\alias{plotNASCByBeamAndLayer}
\title{Plot NASC by Layer and Beam}
\usage{
plotNASCByBeamAndLayer(
  x,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  discardNALayer = TRUE,
  maxDepth = NULL,
  scaleNASC = NULL
)
}
\arguments{
\item{x}{A data.table holding at least the the columns "Beam", "PSU", "MaxLayerDepth" and "NASC".}

\item{xlim, ylim}{The x and y limits of the plot.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{discardNALayer}{Logical: If TRUE discard any NA Layer.}

\item{maxDepth}{If given only the layers down to the maxDepth is included.}
}
\description{
Plot NASC by Layer and Beam
}
