% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{profosPPE2LUF20}
\alias{profosPPE2LUF20}
\title{Convert PROFOS ppe-files to NMDEchosounder or ICESAcoustic files}
\usage{
profosPPE2LUF20(
  ppeFiles,
  acousticFile,
  cruise = "FisherySonar",
  platform = NA,
  acocat = 12,
  freq = 38000,
  transceiver = 1,
  logDistance = 0.1,
  channelThickness = 10,
  minRange = 10,
  transducerDepth = 7,
  depth = "Center.dep",
  upperChannelDepth = 0,
  propellerAngleDegrees = 45,
  checkNAs = TRUE,
  schoolThreshold = c(-70, -20),
  ow = FALSE
)
}
\arguments{
\item{ppeFiles}{A vector of the PROFOS ppe files, or the path to the directory holding the files.}

\item{acousticFile}{The path to the acoustic file to write the vertical distribution from the sonar data to.}
}
\description{
Convert PROFOS ppe-files to NMDEchosounder or ICESAcoustic files
}
