% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SonarVerticalDistribution-package.R
\docType{package}
\name{SonarVerticalDistribution}
\alias{SonarVerticalDistribution}
\alias{SonarVerticalDistribution-package}
\title{Generate NMDechosounder or ICESAcoustic from PROFOS ppe-files}
\description{
Fishery sonar output file in the ppe format from PROFOS are converted to files in the NMDechosounder or ICESAcoustic format suitable for use in StoX.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arnejohannesholmin/SonarVerticalDistribution}
  \item Report bugs at \url{https://github.com/arnejohannesholmin/SonarVerticalDistribution/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@hi.no}

Authors:
\itemize{
  \item Hector Pena
}

Other contributors:
\itemize{
  \item Norwegian Institute of Marine Research [copyright holder, funder]
}

}
